Move webpack config to main.js
This follows the approach as defined in https://storybook.js.org/docs/react/configure/webpack
